# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface HookStep {
  id: string;
  name: string;
  song: string;
  artist: string;
  movieOrAlbum: string;
  tags: string[];
  description: string;
  popularity: number;
  sampleThumbnail: string;
}

export interface YouTubeVideo {
  videoId: string;
  title: string;
  channelTitle: string;
  thumbnailUrl: string;
  publishedAt: string;
  viewCount: number;
  relevanceScore: number;
}

export interface YouTubeSearchResult {
  query: string;
  source: string;
  selected: YouTubeVideo;
  candidates: YouTubeVideo[];
}

export interface ErrorResponse {
  error: string;
}

export type SearchYouTubeParams = {
/**
 * @minLength 1
 */
q: string;
artist?: string;
hookStep?: string;
};