# lib/api-zod/src/generated/types/hookStep.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HookStep {
  id: string;
  name: string;
  song: string;
  artist: string;
  movieOrAlbum: string;
  tags: string[];
  description: string;
  popularity: number;
  sampleThumbnail: string;
}