# lib/api-zod/src/generated/types/youTubeSearchResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { YouTubeVideo } from './youTubeVideo';

export interface YouTubeSearchResult {
  query: string;
  source: string;
  selected: YouTubeVideo;
  candidates: YouTubeVideo[];
}